fix: models bug fixing done.
