+++ src/models/servicesInformation/afterServiceComplaints.js
@@ -1,32 +1,5 @@
 const mongoose = require("mongoose");
-const afterServiceComplaintsStrutureSchema = mongoose.Schema({
-  complaintTitle: {
-    type: String,
-    required: true,
-    minLength: 0,
-    maxLength: 50,
-    unique: true,
-  },
-  complaintDescription: {
-    type: String,
-    required: true,
-    minLength: 0,
-    maxLength: 200,
-  },
-  resolveDescription: {
-    type: String,
-    default: "",
-    minLength: 0,
-    maxLength: 200,
-  },
-  amount: {
-    type: Number,
-    min: 0,
-    max: 25000,
-    required: true,
-    default: 0,
-  },
-});
+const AfterServiceComplaintsStruture = require("../servicesInformation/afterServiceComplaintsStruture");
 const afterServiceComplaintsSchema = mongoose.Schema({
   serviceDataId: {
     type: mongoose.Schema.Types.ObjectId,
@@ -34,7 +7,7 @@ const afterServiceComplaintsSchema = mongoose.Schema({
     required: true,
   },
   complaints: {
-    type: [afterServiceComplaintsStrutureSchema],
+    type: [AfterServiceComplaintsStruture.Schema],
   },
 });
 const AfterServiceComplaints = mongoose.model(
